requirements.txt
discord.py-self
asyncio
requests
urlextract